fix: a lot of minor stuff (WIP)
this changes were discoverd in the process of developing jitsu.

The most WIP thing is the change in ChainMatcher.java with the locking:
it should prevent recursion but is very crude for now.
It should in theory cover the expression = `expression + expression` which leads to an endless cycle of checking if there is an expression before the expression